pyproject.toml
[project]
name = "voice-squads"
version = "0.1.0"
description = ""
authors = [
    {name = "Zefanja Jobse",email = "zefanjaj@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.6.4,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "aiosqlite (>=0.22.1,<0.23.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "environs (>=14.6.0,<15.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
